requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==2.1.1
torch==2.6.0
openai-whisper
